+++ lambda_functions/nci_monitoring/dea_monitoring/es_connection.py
